+++ packages/moonstone/ToggleButton/ToggleButton.js
@@ -30,14 +30,13 @@ const ToggleButtonBase = kind({
 
 	propTypes: /** @lends moonstone/ToggleButton.ToggleButtonBase.prototype */ {
 		/**
-		 * The background-color opacity of this button; valid values are `'opaque'`, `'translucent'`,
-		 * and `'transparent'`.
+		 * The background-color opacity of this button; valid values are `'translucent'`,
+		 * `'lightTranslucent'` and `'transparent'`.
 		 *
 		 * @type {String}
-		 * @default 'opaque'
 		 * @public
 		 */
-		backgroundOpacity: PropTypes.oneOf(['opaque', 'translucent', 'transparent']),
+		backgroundOpacity: PropTypes.oneOf(['translucent', 'lightTranslucent', 'transparent']),
 
 		/**
 		 * The string to be displayed as the main content of the toggle button.
@@ -121,7 +120,6 @@ const ToggleButtonBase = kind({
 	},
 
 	defaultProps: {
-		backgroundOpacity: 'opaque',
 		disabled: false,
 		minWidth: true,
 		selected: false,